# optimize(read-only): Add flags to optimize read-only operations