--- a/pkg/uefi/file.go
+++ b/pkg/uefi/file.go
@@ -436,10 +436,15 @@ func NewFile(buf []byte) (*File, error) {
 		return nil, fmt.Errorf("File size too big! File with GUID: %v has length %v, but is only %v bytes big",
 			f.Header.GUID, f.Header.ExtendedSize, buflen)
 	}
-	// Copy out the buffer.
-	newBuf := buf[:f.Header.ExtendedSize]
-	f.buf = make([]byte, f.Header.ExtendedSize)
-	copy(f.buf, newBuf)
+
+	if ReadOnly {
+		f.buf = buf[:f.Header.ExtendedSize]
+	} else {
+		// Copy out the buffer.
+		newBuf := buf[:f.Header.ExtendedSize]
+		f.buf = make([]byte, f.Header.ExtendedSize)
+		copy(f.buf, newBuf)
+	}
 
 	// Special case for NVAR Store stored in raw file
 	if f.Header.Type == FVFileTypeRaw && f.Header.GUID == *NVAR {