--- a/pkg/uefi/firmwarevolume.go
+++ b/pkg/uefi/firmwarevolume.go
@@ -259,10 +259,14 @@ func NewFirmwareVolume(data []byte, fvOffset uint64, resizable bool) (*FirmwareV
 	fv.FVType = FVGUIDs[fv.FileSystemGUID]
 	fv.FVOffset = fvOffset
 
-	// copy out the buffer.
-	newBuf := data[:fv.Length]
-	fv.buf = make([]byte, fv.Length)
-	copy(fv.buf, newBuf)
+	if ReadOnly {
+		fv.buf = data[:fv.Length]
+	} else {
+		// copy out the buffer.
+		newBuf := data[:fv.Length]
+		fv.buf = make([]byte, fv.Length)
+		copy(fv.buf, newBuf)
+	}
 
 	// Parse the files.
 	// TODO: handle fv data alignment.