We use fiano to parse UEFI images and we do not need to modify anything or/and decompress, but buffer copies and decompression consumes >90% CPU and a lot of RAM, therefore we add flags to make that disablable.

Predecessor-PR: https://github.com/linuxboot/fiano/pull/305

---

### Benchmark the `uefi.Parse` on EDKII image GALAGOPRO3

Before (with disabled Flags):
```
xaionaro@void:~/go/src/github.com/facebookexternal/pcr0tool/pkg/uefi$ GO111MODULE=off go test ./ -bench=. -benchtime=10s -memprofile=/tmp/heap.pprof 2>&1 | grep -v warning
goos: linux
goarch: amd64
pkg: github.com/facebookexternal/pcr0tool/pkg/uefi
      67         208900801 ns/op        171524652 B/op    856601 allocs/op
PASS
ok      github.com/facebookexternal/pcr0tool/pkg/uefi   15.146s
```

After (with enabled Flags):
```
xaionaro@void:~/go/src/github.com/facebookexternal/pcr0tool/pkg/uefi$ GO111MODULE=off go test ./ -bench=. -benchtime=10s -memprofile=/tmp/heap.pprof 2>&1 | grep -v warning
goos: linux
goarch: amd64
pkg: github.com/facebookexternal/pcr0tool/pkg/uefi
   15480            817542 ns/op           81789 B/op       2594 allocs/op
PASS
ok      github.com/facebookexternal/pcr0tool/pkg/uefi   22.521s
```

---

### Executing our tool one a random sample

Before (with disabled Flags):
```
        User time (seconds): 1.12
        System time (seconds): 0.02
        Percent of CPU this job got: 104%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 0:01.10
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 272860
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 0
        Minor (reclaiming a frame) page faults: 10466
        Voluntary context switches: 1002
        Involuntary context switches: 72
        Swaps: 0
        File system inputs: 0
        File system outputs: 0
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```

After (with enabled Flags):
```
        User time (seconds): 0.11
        System time (seconds): 0.01
        Percent of CPU this job got: 106%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 0:00.12
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 109988
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 0
        Minor (reclaiming a frame) page faults: 5233
        Voluntary context switches: 186
        Involuntary context switches: 1
        Swaps: 0
        File system inputs: 0
        File system outputs: 0
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```