@xaionaro 2021-02-06
> That's too high risk for a .9 second performance gain.

I would say that it is 250x performance gain:
```
208900801 / 817542
255.52302022403741948425
```

It means that without this optimization I will be have to use 25000 of servers instead of 100 servers. Which is quite critical.

> Until you can leave this option always enabled, I'm opposed to making this change. If someone uses the flag and creates a bad image, it can brick a machine.

Well. This is disappointing. But I respect the decision, of course.