@xaionaro 2021-02-08
> I will approve again if you can get it to pass tests.

Thank you a lot! But still I'm wondering: would you prefer me to the intermediate structure unsafeOptions to hide the options or to leave as is?