@xaionaro 2021-02-09
I just made a warning more explicit:
```
	// DO NOT USE THIS OPTION UNLESS YOU KNOW WHAT ARE YOU DOING. IF YOU
	// WILL MODIFY A FIRMWARE WITH THIS OPTION BEING ENABLED, THIS FIRMWARE
	// MIGHT BRICK YOUR DEVICE.
```
And approves were again reset. Sorry. Could you re-approve one last time? :(